// src/modules/merchandising/reports/report/date-wise-shiplent-summary-report/date-wise-shiplent-summary-report-type.ts
export type DateWiseShiplentSummaryReportType = {
  FACTORYID?: string;
  CHALLANDATE: string;
  BUYERID?: number;
  STYLE_ID?: number;
  PONO?: string;
  ORDER_QTY: number;
  FOB: number;
  COMPANY_NAME?: string;
  COMPANY_ADDRESS?: string;
  BUYER?: string;
  STYLENAME?: string;
  STYLENO?: string;
  ITEMUOM?: string;
  SHIPMENTQTY: number;
  PREV_SHIPMENTQTY: number;
};